Consolidate threading utility tests into one file
