--- tests/test_named_locks.py
@@ -1,55 +1,0 @@
-# This file is part of daf_butler.
-#
-# Developed for the LSST Data Management System.
-# This product includes software developed by the LSST Project
-# (http://www.lsst.org).
-# See the COPYRIGHT file at the top-level directory of this distribution
-# for details of code ownership.
-#
-# This software is dual licensed under the GNU General Public License and also
-# under a 3-clause BSD license. Recipients may choose which of these licenses
-# to use; please see the files gpl-3.0.txt and/or bsd_license.txt,
-# respectively.  If you choose the GPL option then the following text applies
-# (but note that there is still no warranty even if you opt for BSD instead):
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-import unittest
-
-from lsst.daf.butler._utilities.named_locks import NamedLocks
-
-
-class NamedLocksTestCase(unittest.TestCase):
-    """Test NamedLocks."""
-
-    def test_named_locks(self):
-        locks = NamedLocks()
-        lock1 = locks._get_lock("a")
-        lock2 = locks._get_lock("b")
-        lock3 = locks._get_lock("a")
-
-        self.assertIs(lock1, lock3)
-        self.assertIsNot(lock1, lock2)
-
-        self.assertFalse(lock1.locked())
-        self.assertFalse(lock2.locked())
-        with locks.lock("a"):
-            self.assertTrue(lock1.locked())
-            self.assertFalse(lock2.locked())
-        self.assertFalse(lock1.locked())
-        self.assertFalse(lock2.locked())
-
-
-if __name__ == "__main__":
-    unittest.main()
